ObjectPath classname must not be nil
ObjectPath.new with a classname of nil will pass a NULL classname to
sfcc newObjectPath. However, sfcc's ObjectPath.toString will crash on
a NULL classname.

This patch makes sure that a missing classname to ObjectPath.new will
be translated to an empty string.
An explicit nil classname to ObjectPath.new will raise an ArgumentError.